Move the test cases to src/
