+++ phpunit.xml.dist
@@ -16,7 +16,7 @@
 
     <testsuites>
         <testsuite name="unit">
-            <directory>./tests/</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
 
@@ -24,8 +24,9 @@
         <whitelist>
             <directory>src</directory>
             <exclude>
-                <directory>src/Moka/Traits</directory>
                 <directory>src/Moka/Exception</directory>
+                <directory>src/Moka/Tests</directory>
+                <directory>src/Moka/Traits</directory>
             </exclude>
         </whitelist>
     </filter>
