+++ tests/Plugin/Phake/PhakeMockingStrategyTest.php
@@ -6,12 +6,12 @@ namespace Tests\Plugin\Phake;
 use Moka\Exception\MockNotCreatedException;
 use Moka\Factory\StubFactory;
 use Moka\Plugin\Phake\PhakeMockingStrategy;
+use Moka\Tests\MokaMockingStrategyTestCase;
 use Tests\BarTestClass;
 use Tests\FooTestClass;
-use Tests\Strategy\MockingStrategyTestCase;
 use Tests\TestInterface;
 
-class PhakeMockingStrategyTest extends MockingStrategyTestCase
+class PhakeMockingStrategyTest extends MokaMockingStrategyTestCase
 {
     public function __construct($name = null, array $data = [], $dataName = '')
     {
